Feature count displaying

+++ src/map/renderlayer.h
@@ -45,7 +45,7 @@ public:
     virtual void draw(enum ngsDrawState state, OGREnvelope extent, double zoom,
                         float level);
     virtual float getComplete() const;
-    virtual size_t getFeatureCount() const;
+    virtual int getFeatureCount() const;
 protected:
     virtual void clearTiles() = 0;
     virtual void drawTiles() = 0;
@@ -59,7 +59,7 @@ protected:
     unsigned char m_renderZoom;
     float m_renderLevel;
     float m_complete;
-    size_t m_featureCount;
+    int m_featureCount;
     MapView* m_mapView;
     CPLLock *m_hThreadLock;
     StyleUPtr m_style;
